database tables and abstractions
Here we start creating some abstractions, to be able to manage the
shopping list with different implementations (e.g. a memory based and
a database based one).

We now have a maker function that returns a function for managing the
data. The functions takes in an action and optionally some
payload (shopping item), and then induce a side effect in order to
update, add or remove the data. It can also return data: get the
shopping list or query a specific subset of the list (which is not
done yet).